feat: simplify URL param logic
